+++ rest_framework_api_key/models.py
@@ -1,29 +1,36 @@
 from typing import Tuple
+
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils import timezone
-from ._helpers import check_key, generate_key
+
+from .crypto import KeyGenerator
 
 
-class APIKeyManager(models.Manager):
-    def create_key(self, **kwargs) -> Tuple["APIKey", str]:
+class BaseAPIKeyManager(models.Manager):
+    key_generator = KeyGenerator()
+
+    def assign_key(self, obj: "AbstractAPIKey") -> str:
+        key, hashed_key = self.key_generator.generate()
+        obj.id = hashed_key
+        return key
+
+    def create_key(self, **kwargs) -> Tuple["AbstractAPIKey", str]:
         # Prevent from manually setting the primary key.
         kwargs.pop("id", None)
-
-        obj = self.model(**kwargs)  # type: APIKey
-
-        generated_key, key_id = generate_key()
-        obj.id = key_id
+        obj = self.model(**kwargs)  # type: AbstractAPIKey
+        key = self.assign_key(obj)
         obj.save()
-
-        return obj, generated_key
+        return obj, key
 
     def is_valid(self, key: str) -> bool:
         prefix, _, _ = key.partition(".")
 
         try:
-            api_key = self.get(id__startswith=prefix, revoked=False)
-        except self.model.DoesNotExist:
+            api_key = self.get(
+                id__startswith=prefix, revoked=False
+            )  # type: AbstractAPIKey
+        except (self.model.DoesNotExist, self.model.MultipleObjectsReturned):
             return False
 
         if not api_key.is_valid(key):
@@ -35,7 +42,11 @@ class APIKeyManager(models.Manager):
         return True
 
 
-class APIKey(models.Model):
+class APIKeyManager(BaseAPIKeyManager):
+    pass
+
+
+class AbstractAPIKey(models.Model):
     objects = APIKeyManager()
 
     id = models.CharField(max_length=100, unique=True, primary_key=True)
@@ -66,6 +77,7 @@ class APIKey(models.Model):
     )
 
     class Meta:  # noqa
+        abstract = True
         ordering = ("-created",)
         verbose_name = "API key"
         verbose_name_plural = "API keys"
@@ -92,7 +104,7 @@ class APIKey(models.Model):
 
     def is_valid(self, key: str) -> bool:
         _, _, hashed_key = self.pk.partition(".")
-        return check_key(key, hashed_key)
+        return type(self).objects.key_generator.verify(key, hashed_key)
 
     def clean(self):
         self._validate_revoked()
@@ -109,3 +121,7 @@ class APIKey(models.Model):
 
     def __str__(self) -> str:
         return str(self.name)
+
+
+class APIKey(AbstractAPIKey):
+    pass