+++ tests/conftest.py
@@ -21,6 +21,7 @@ def pytest_configure():
                 # Project
                 "rest_framework",
                 "rest_framework_api_key",
+                "tests.heroes",
             ],
             TEMPLATES=[
                 # Admin